# src/main/java/com/dsw01/practica02/dto/DepartamentoUpdateRequest.java
package com.dsw01.practica02.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record DepartamentoUpdateRequest(
        @NotBlank(message = "nombre es obligatorio")
        @Size(max = 100, message = "nombre debe tener maximo 100 caracteres")
        String nombre
) {
}